+++ cilantro/include/graphics/IFramebuffer.h
@@ -13,7 +13,7 @@ struct IFramebuffer
 
     virtual void BindFramebuffer () const = 0;
     virtual void BindFramebufferColorTexturesAsColor () const = 0;
-    virtual void BindFramebufferDepthArrayTextureAsColor (size_t index) const = 0;
+    virtual void BindFramebufferDepthArrayTextureAsColor (unsigned int index) const = 0;
     virtual void BindFramebufferDepthArrayTextureAsDepth () const = 0;
     virtual void BindFramebufferRenderbuffer () const = 0;
     
@@ -21,16 +21,16 @@ struct IFramebuffer
 
     virtual void BlitFramebuffer () const = 0;
 
-    virtual uint32_t GetWidth () const = 0;
-    virtual uint32_t GetHeight () const = 0;
+    virtual unsigned int GetWidth () const = 0;
+    virtual unsigned int GetHeight () const = 0;
 
-    virtual size_t GetColorTextureCount() const = 0;
-    virtual size_t GetRGBTextureCount () const = 0;
-    virtual size_t GetRGBATextureCount () const = 0;
-    virtual size_t GetDepthArrayLayerCount () const = 0;
+    virtual unsigned int GetColorTextureCount() const = 0;
+    virtual unsigned int GetRGBTextureCount () const = 0;
+    virtual unsigned int GetRGBATextureCount () const = 0;
+    virtual unsigned int GetDepthArrayLayerCount () const = 0;
 
     virtual bool IsDepthStencilRenderbufferEnabled () const = 0;
     virtual bool IsDepthArrayEnabled () const = 0;
 
-    virtual void SetFramebufferResolution (uint32_t bufferWidth, uint32_t bufferHeight) = 0;
+    virtual void SetFramebufferResolution (unsigned int bufferWidth, unsigned int bufferHeight) = 0;
 };
